Add menu action and better component split
