+++ webapp/src/assets/locales/fr.json
@@ -227,6 +227,7 @@
       "totalPlayer": "Joueurs prêts",
       "tryNumber": "Nombre d'essais"
     },
+    "issue": "Si vous rencontrez le moindre problème, rejoignez notre",
     "leave": "Quitter la session",
     "name": {
       "0": "Le Navire Fantôme",
@@ -332,6 +333,8 @@
     },
     "player": {
       "notReady": "Pas prêt",
+      "online": "Joueur connecté",
+      "onlines": "Joueurs connectés",
       "ready": "Prêt",
       "status": {
         "closed": "Hors ligne",
@@ -340,8 +343,9 @@
         "started": "Jeu lancé"
       }
     },
-    "playersLabel": "Joueurs",
     "playerLabel": "Joueur",
+    "playersLabel": "Joueurs",
+    "refresh": "Raffraichir",
     "run": "Lever l'ancre",
     "status": {
       "countdown": "Lancement dans",
